Add support to skip key encryption
In some cases it is not desirable to encrypt the private key when
generating a certificate. For example, I want to use this module to
generate a test certificate for Cockpit, which does not use encrypted
keys.

To accomodate this I add the new option `encryptKey` which is set to
`true` by default to ensure backward compatability. As this value is
a boolean I could not use `Object.assign` to assign this value, but
normal old-fashioned assignment works fine.